# apps/backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Applad API
  version: 1.0.0
  description: Self-hosted Backend-as-a-Service API
  license:
    name: BSD-3-Clause
servers:
  - url: /v1
    description: Default (relative)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-Applad-Key
  parameters:
    projectHeader:
      name: X-Applad-Project
      in: header
      required: true
      schema:
        type: string
...
paths:
...
  /edge/functions:
    post:
      summary: Create an edge function
      tags: [Edge]
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/projectHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                source:
                  type: string
      responses:
        '201':
          description: Edge function created
    get:
      summary: List edge functions
      tags: [Edge]
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/projectHeader'
      responses:
        '200':
          description: Edge function list
  /edge/functions/{functionId}:
    get:
      summary: Get an edge function
      tags: [Edge]
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/projectHeader'
        - name: functionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Edge function details
    put:
      summary: Update an edge function
      tags: [Edge]
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/projectHeader'
        - name: functionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete an edge function
      tags: [Edge]
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/projectHeader'
        - name: functionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
...
  /billing/plans:
    get:
      summary: List available plans
      tags: [Billing]
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/projectHeader'
      responses:
        '200':
          description: Plan list
  /billing/subscription:
    get:
      summary: Get current subscription
      tags: [Billing]
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/projectHeader'
      responses:
        '200':
          description: Subscription details
    post:
      summary: Subscribe to a plan
      tags: [Billing]
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/projectHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [planId]
              properties:
                planId:
                  type: string
      responses:
        '201':
          description: Subscribed
    delete:
      summary: Cancel subscription
      tags: [Billing]
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/projectHeader'
      responses:
        '204':
          description: Cancelled
  /billing/usage:
    get:
      summary: Get usage summary
      tags: [Billing]
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/projectHeader'
      responses:
        '200':
          description: Usage summary
  /billing/invoices:
    get:
      summary: List invoices
      tags: [Billing]
      security:
        - bearerAuth: []
        - apiKey: []
      parameters:
        - $ref: '#/components/parameters/projectHeader'
      responses:
        '200':
          description: Invoice list

  # ── Regions ────────────────────────────────────────────────────────────────
...